requirements.txt
openai>=1.30.0
Pillow>=10.2.0
pdf2image>=1.17.0
